requirements.txt
Django==4.1
django_guardian==2.4.0
